backend/pyproject.toml
[project]
name = "ai-land"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "langchain[openai]>=0.3.25",
    "langchain-openai>=0.3.19",
    "langgraph>=0.4.8",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
